+++ MockSwiftExample/MockSwiftExample/Basics.swift
@@ -1,4 +1,4 @@
-//Basic.swift
+//Basics.swift
 /*
  MIT License
  
@@ -25,11 +25,12 @@
 
 import Foundation
 
-protocol Basic {
+//sourcery: AutoMockable
+protocol Basics {
   func doSomething()
   func doSomething() -> Int
   func doSomething() -> String?
-  func doSomething(arg: String) -> String
+  func doSomething(arg: String) throws -> String
   func doSomething(arg1: String, arg2: Int?) -> String
   func doSomething(with arg: String) -> String
   func doSomething(with arg1: String, and arg2: Bool) -> String